+++ packages/veui/src/components/Select/Select.vue
@@ -38,7 +38,10 @@
               :selected="subOption.value === value"
               :key="subOption.value"
               @select="handleSelect(subOption)">
-              <slot name="option" v-bind="subOption" :selected="option.value === value"></slot>
+              <slot v-if="$scopedSlots.option" name="option" v-bind="subOption" :selected="option.value === value"></slot>
+              <template v-if="$scopedSlots['option-label']" slot="label">
+                <slot name="option-label" v-bind="subOption" :selected="option.value === value"></slot>
+              </template>
             </veui-option>
           </div>
           <veui-option
@@ -48,7 +51,10 @@
             :selected="option.value === value"
             :key="option.value"
             @select="handleSelect(option)">
-            <slot name="option" v-bind="option" :selected="option.value === value"></slot>
+            <slot v-if="$scopedSlots.option" name="option" v-bind="option" :selected="option.value === value"></slot>
+            <template v-if="$scopedSlots['option-label']" slot="label">
+              <slot name="option-label" v-bind="option" :selected="option.value === value"></slot>
+            </template>
           </veui-option>
         </template>
       </slot>